Added dynamic memory management for Loci. Use command option --dmm to activate it. And also if use --dmm, use --dmmverbose to print out some info. about the memory management, use --decoration to visualize the decorated multilevel graph, provide that the GraphViz package(from AT&T research) installed on the path. When not using --dmm, these options are disabled.
visitor.h and visitor.cc are newly added files. They are the main implementation of the memory management. Also the constuction and compilation of various compilers for the multilevel graph has been seperated. The memory management scheme uses the visitor architecture as a basic strategy and decorating the multilevel graph to include memory specific information.