This commit adds an "erase" method into the contrainer classes. It is intended to be used to "shrink" a container by removing part of its domain. Currently only the DMap has it implemented, others all default to do nothing.
Two Loci bugs have also been fixed. One is in the "ORBPartition" method in "orb.cc" file. In the sequential case, it didn't set the item to process map correctly. The other bug is in the "intervalSet::Union(interval)" method. It does not merge intervals correctly under certain rare circumstances.