MAH[#4]: Merge snapshot into dev
We need to go ahead and merge snapshot into dev. Since we (CFD) are
actively developing on the `dev` branch and Ed is developing on the
`snapthot`, all bug fixes that we receive from Ed are based off the
snapshot development. It does not makes sense anymore for the two
parties to be developing on two diverging branches.

---------

Co-authored-by: Qiuhan Arnoldus <qiuhanxue@gmail.com>
Co-authored-by: Ed Luke <edluke@gmail.com>
Co-authored-by: cvs2git <>
Co-authored-by: Mark Hunt <mark.hunt@cfd-research.com>